Use gettimeofday for TIME_get_ms(). monotonic_seconds() is a realtime timer that could resort to tsc.

+++ crashreporter/crashreporter.cpp
@@ -427,7 +427,7 @@ static double get_s(void) {
     if (gettimeofday(&now, NULL) != 0)
       return 0.0;
 
-    return now.tv_sec + now.tv_usec/(1000.0*1000.0);
+    return now.tv_sec + (double)now.tv_usec/(1000.0*1000.0);
 }
